% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.gibbs.R
\name{hmm.gibbs}
\alias{hmm.gibbs}
\title{Parameter Update at each Gibbs iteration}
\usage{
hmm.gibbs(y, var, p_matrix, tran_prior, var_prior, q, s.out = FALSE)
}
\arguments{
\item{y}{Vector of mixture Gaussian observation.}

\item{var}{List of state variance.}

\item{p_matrix}{Transition matrix with \verb{(i,j) }entry being transition probability from state i to state j.}

\item{tran_prior}{List of parameters for transition probability priors(See \code{\link{transition}}).}

\item{var_prior}{List of parameters for prior of state variance(See \code{\link{rIG}}).}

\item{q}{Scaler, the period on which Variance Ratio Test is based}

\item{s.out}{Logical; if TRUE, it will return the latent state values}
}
\value{
Either a \verb{nsamples x 2} matrix of VR statistcs with each row corresponding to an MCMC update cycle, or a list with element VR and possibly \verb{nsamples x 4} matrix of Theta and \verb{nsamples x \code{length(y)}} matrix of S, if the parameter posterior and the latent state values are required
}
\description{
Calculate statistics reuqired for gibbs-sampling-augmented variance ratio test, the latent State series of the Markove Switching model, posterior parameter values and VR statistics for a single iteration of Gibbs sampler.
}
