% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIG.R
\name{rIG}
\alias{rIG}
\title{Ramdom sampling from the Inverse-Gamma distribution}
\usage{
rIG(n = 1, y, S, prior_par = list(shape = 1/2, rate = 1/2))
}
\arguments{
\item{n}{Numer of random draws, default to be 1. If \code{length(n)>1}, the length is taken to be the number required.}

\item{y}{Vector of mixture Gaussian observation.}

\item{S}{Vector of the state values. Each element of S must be either 1 or 2 indicating the state value.}

\item{prior_par}{List of shape and rate parameter values of priors.}
}
\value{
List of posterior variance for each state.
}
\description{
Ramdom sampling from the Inverse-Gamma distribution
}
\details{
The Inverse-Gamma distribution here with shape parameter \code{a} and scale parameter \code{b} has PDF\preformatted{f(x | shape = a, rate = b) = cst * x^(a + 1) * exp(-b/x),   cst = b^a/Gamma(a).
}

The distribution is proper only when \verb{a,b > 0}, but these values are unrestricted here since improper priors can be useful for Bayesian inference.  When either \verb{a,b <= 0} the function sets \code{cst = 1} in the expression above.

The function takes advantage of the fact that if X ~ Gamma(a,b), then 1/X ~ IG(a,b).
}
\seealso{
\code{\link{rTIG}}
}
