\name{targeted-selection-package}
\alias{targeted-selection-package}
\alias{targeted-selection}
\docType{package}
\title{
\packageTitle{targeted-selection}
}
\description{
\packageDescription{targeted-selection}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{targeted-selection}
\packageIndices{targeted-selection}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{targeted-selection}

Maintainer: \packageMaintainer{targeted-selection}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
