% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMS-package.R
\docType{package}
\name{TMS-package}
\alias{TMS}
\alias{TMS-package}
\title{TMS: Bayesian Inference for Two-State Markov Switching Model}
\description{
Parameter inference of two-sate Markov switching model of mixture Gaussian distributio. The package also provide tools for Gibbs-sampling-augmented Variance Ratio test.
}
\author{
\strong{Maintainer}: Ye Meng \email{ye.meng@uwaterloo.ca}

}
\keyword{internal}
