% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tms.fit.R
\name{tms.fit}
\alias{tms.fit}
\title{Posterior Inferencefor Two-State Markov Switching model}
\usage{
tms.fit(
  y,
  q,
  burn,
  nsamples,
  s.out = FALSE,
  par.out = FALSE,
  tran_prior = NULL,
  var_prior = NULL
)
}
\arguments{
\item{y}{Vector of mixture Gaussian observation.}

\item{q}{Scaler, the period on which Variance Ratio Test is based}

\item{burn}{Integer number of initial iterations to discard as burn-in. The default value is \code{min(nsamples/10, 1000)}.}

\item{nsamples}{Number of posterior iterations (not including burn-in)}

\item{s.out}{Logical; if TRUE, it will return the latent state values}

\item{par.out}{Logical, if TRUE, it will return the posterior draws for parameters}

\item{tran_prior}{List of parameters for prior of transition probabilities(See \code{\link{transition}})}

\item{var_prior}{List of parameters for prior of state variance(See \code{\link{rIG}}).}
}
\value{
\verb{nsampls x 2} matrix of variance ratio test of standardized observations and randomized observations.
}
\description{
Calculate posterior variance ratio statistics of standardized observations and randomized observations.
}
\details{
The Gibbs Sampler will update latent state values, state variance and transition matrix in order.
}
\seealso{
\code{\link{hmm.gibbs}} for update at each iteration
}
