% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{tms}
\alias{tms}
\title{Targeted model selection}
\usage{
tms(f1, f2, df, R = 200)
}
\arguments{
\item{f1}{An asymptotically unbiased estimator of the parameter of interest}

\item{f2}{A potentially biased estimator of the parameter of interest}

\item{df}{The data}

\item{R}{Number of bootstrap draws}
}
\description{
Targeted model selection
}
