% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.num.R
\name{trans.num}
\alias{trans.num}
\title{State Transition Steps for Two-Satet Makov Chian}
\usage{
trans.num(S)
}
\arguments{
\item{S}{Vector of the state values. Each element of S must be either 1 or 2 indicating the state values.}
}
\value{
A list of the number of transitions of all four possible transition.
}
\description{
Count the transitions happend in S.
}
