% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VR.pval.R
\name{VR.pval}
\alias{VR.pval}
\title{P-Values for Bayesian Variance Ratio Test}
\usage{
VR.pval(vr_t)
}
\arguments{
\item{vr_t}{Matrix with each row being the result from each sweep of Gibbs sampler}
}
\value{
p-value of variance ratio test.
}
\description{
Calculate p-values for statistics based on Bayesian Inference for variance ratio test.
}
\details{
After Gibbs sampler, the p-value is given by\preformatted{#(VR>VR_star)/M
}
}
