% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VR.stat.R
\name{VR.stat}
\alias{VR.stat}
\title{Singular Variance Ratio Statistics}
\usage{
VR.stat(y_sd, q)
}
\arguments{
\item{y_sd}{Vector of standardized observations}

\item{q}{Scaler, the period on which Variance Ratio Test is based}
}
\value{
q-period variance ratio statistics
}
\description{
Calculate variance ratio given series y to test random-walk hypothesis
}
\details{
If y_sd follow a random walk, the returned value should be 1.
}
