% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VR.test.R
\name{VR.test}
\alias{VR.test}
\title{Bayesian Statistics for Variance Ratio Test}
\usage{
VR.test(y, sigma, q)
}
\arguments{
\item{y}{Vector of mixture Gaussian observation.}

\item{sigma}{Vector of standard deviation at each state.Same length as y.}

\item{q}{Scalar, the period on which Variance Ratio Test is based}
}
\value{
List of variance ratio statistic of y and randomized realizations.
}
\description{
Calculate variance ratio statistics for mixture Gaussian observations and its randomized realizations.
}
\details{
To keep the information contained in sigma, the randomization will be a product of sigma and random draws from standard normal distribution.
}
