\name{tms-package}
\alias{tms-package}
\alias{tms}
\docType{package}
\title{
\packageTitle{tms}
}
\description{
\packageDescription{tms}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tms}
\packageIndices{tms}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tms}

Maintainer: \packageMaintainer{tms}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
